refactor: move `num_blocks_per_session` param to shared module
