+++ e2e/tests/update_params.feature
@@ -34,14 +34,14 @@ Feature: Params Namespace
 
   # NB: If you are reading this and the proof module has parameters
   # that are not being updated in this test, please update the test.
-  Scenario: An authorized user updates all "session" module params
+  Scenario: An authorized user updates all "shared" module params
     Given the user has the pocketd binary installed
-    And all "session" module params are set to their default values
-    And an authz grant from the "gov" "module" account to the "pnf" "user" account for the "/poktroll.session.MsgUpdateParams" message exists
-    When the "pnf" account sends an authz exec message to update all "session" module params
+    And all "shared" module params are set to their default values
+    And an authz grant from the "gov" "module" account to the "pnf" "user" account for the "/poktroll.shared.MsgUpdateParams" message exists
+    When the "pnf" account sends an authz exec message to update all "shared" module params
       | name                   | value | type  |
       | num_blocks_per_session | 10    | int64 |
-    Then all "session" module params should be updated
+    Then all "shared" module params should be updated
 
   # NB: If you are reading this and any module has parameters that
   # are not being updated in this test, please update the test.
