+++ proto/poktroll/session/params.proto
@@ -10,6 +10,4 @@ import "gogoproto/gogo.proto";
 message Params {
   option (amino.name) = "poktroll/x/session/Params";
   option (gogoproto.equal) = true;
-
-  uint64 num_blocks_per_session = 1 [(gogoproto.jsontag) = "num_blocks_per_session"];
 }
