+++ x/session/keeper/msg_update_params.go
@@ -10,6 +10,7 @@ func (k msgServer) UpdateParams(ctx context.Context, req *types.MsgUpdateParams)
 	if err := req.ValidateBasic(); err != nil {
 		return nil, err
 	}
+
 	if k.GetAuthority() != req.Authority {
 		return nil, types.ErrSessionInvalidSigner.Wrapf("invalid authority; expected %s, got %s", k.GetAuthority(), req.Authority)
 	}
