+++ x/session/types/errors.go
@@ -15,6 +15,4 @@ var (
 	ErrSessionInvalidService         = sdkerrors.Register(ModuleName, 1106, "invalid service in session")
 	ErrSessionInvalidBlockHeight     = sdkerrors.Register(ModuleName, 1107, "invalid block height for session")
 	ErrSessionInvalidSessionId       = sdkerrors.Register(ModuleName, 1108, "invalid sessionId")
-	ErrSessionParamNameInvalid       = sdkerrors.Register(ModuleName, 1109, "the provided param name is invalid")
-	ErrSessionParamInvalid           = sdkerrors.Register(ModuleName, 1110, "the provided param is invalid")
 )
