+++ x/session/types/params.go
@@ -2,15 +2,7 @@ package types
 
 import paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
 
-const (
-	DefaultNumBlocksPerSession = 4
-	ParamNumBlocksPerSession   = "num_blocks_per_session"
-)
-
-var (
-	_                      paramtypes.ParamSet = (*Params)(nil)
-	KeyNumBlocksPerSession                     = []byte("NumBlocksPerSession")
-)
+var _ paramtypes.ParamSet = (*Params)(nil)
 
 // ParamKeyTable the param key table for launch module
 func ParamKeyTable() paramtypes.KeyTable {
@@ -19,9 +11,7 @@ func ParamKeyTable() paramtypes.KeyTable {
 
 // NewParams creates a new Params instance
 func NewParams() Params {
-	return Params{
-		NumBlocksPerSession: DefaultNumBlocksPerSession,
-	}
+	return Params{}
 }
 
 // DefaultParams returns a default set of parameters
@@ -30,36 +20,11 @@ func DefaultParams() Params {
 }
 
 // ParamSetPairs get the params.ParamSet
-func (params *Params) ParamSetPairs() paramtypes.ParamSetPairs {
-	return paramtypes.ParamSetPairs{
-		paramtypes.NewParamSetPair(
-			KeyNumBlocksPerSession,
-			&params.NumBlocksPerSession,
-			ValidateNumBlocksPerSession,
-		),
-	}
+func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
+	return paramtypes.ParamSetPairs{}
 }
 
 // ValidateBasic does a sanity check on the provided params.
-func (params *Params) ValidateBasic() error {
-	if err := ValidateNumBlocksPerSession(params.NumBlocksPerSession); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// ValidateNumBlocksPerSession validates the NumBlocksPerSession param
-// NB: The argument is an interface type to satisfy the ParamSetPair function signature.
-func ValidateNumBlocksPerSession(v interface{}) error {
-	numBlocksPerSession, ok := v.(uint64)
-	if !ok {
-		return ErrSessionParamInvalid.Wrapf("invalid parameter type: %T", v)
-	}
-
-	if numBlocksPerSession < 1 {
-		return ErrSessionParamInvalid.Wrapf("invalid NumBlocksPerSession: (%v)", numBlocksPerSession)
-	}
-
+func (p Params) ValidateBasic() error {
 	return nil
 }
