+++ x/shared/types/errors.go
@@ -8,6 +8,7 @@ import (
 
 // x/shared module sentinel errors
 var (
-	ErrInvalidSigner = sdkerrors.Register(ModuleName, 1100, "expected gov account as only signer for proposal message")
-	ErrSample        = sdkerrors.Register(ModuleName, 1101, "sample error")
+	ErrInvalidSigner           = sdkerrors.Register(ModuleName, 1100, "expected gov account as only signer for proposal message")
+	ErrSessionParamNameInvalid = sdkerrors.Register(ModuleName, 1101, "the provided param name is invalid")
+	ErrSessionParamInvalid     = sdkerrors.Register(ModuleName, 1102, "the provided param is invalid")
 )
