+++ x/shared/types/params.go
@@ -4,7 +4,15 @@ import (
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
 )
 
-var _ paramtypes.ParamSet = (*Params)(nil)
+const (
+	DefaultNumBlocksPerSession = 4
+	ParamNumBlocksPerSession   = "num_blocks_per_session"
+)
+
+var (
+	_                      paramtypes.ParamSet = (*Params)(nil)
+	KeyNumBlocksPerSession                     = []byte("NumBlocksPerSession")
+)
 
 // ParamKeyTable the param key table for launch module
 func ParamKeyTable() paramtypes.KeyTable {
@@ -13,7 +21,9 @@ func ParamKeyTable() paramtypes.KeyTable {
 
 // NewParams creates a new Params instance
 func NewParams() Params {
-	return Params{}
+	return Params{
+		NumBlocksPerSession: DefaultNumBlocksPerSession,
+	}
 }
 
 // DefaultParams returns a default set of parameters
@@ -22,11 +32,36 @@ func DefaultParams() Params {
 }
 
 // ParamSetPairs get the params.ParamSet
-func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
-	return paramtypes.ParamSetPairs{}
+func (params *Params) ParamSetPairs() paramtypes.ParamSetPairs {
+	return paramtypes.ParamSetPairs{
+		paramtypes.NewParamSetPair(
+			KeyNumBlocksPerSession,
+			&params.NumBlocksPerSession,
+			ValidateNumBlocksPerSession,
+		),
+	}
 }
 
 // Validate validates the set of params
-func (p Params) Validate() error {
+func (params *Params) ValidateBasic() error {
+	if err := ValidateNumBlocksPerSession(params.NumBlocksPerSession); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// ValidateNumBlocksPerSession validates the NumBlocksPerSession param
+// NB: The argument is an interface type to satisfy the ParamSetPair function signature.
+func ValidateNumBlocksPerSession(v interface{}) error {
+	numBlocksPerSession, ok := v.(uint64)
+	if !ok {
+		return ErrSessionParamInvalid.Wrapf("invalid parameter type: %T", v)
+	}
+
+	if numBlocksPerSession < 1 {
+		return ErrSessionParamInvalid.Wrapf("invalid NumBlocksPerSession: (%v)", numBlocksPerSession)
+	}
+
 	return nil
 }
